// src/types/error.ts
export interface AppError {
  title: string;
  message: string;
}

export enum AppErrorCode {
  // Authentication errors
  AUTH_EMAIL_ALREADY_EXISTS = "AUTH_EMAIL_ALREADY_EXISTS",
  AUTH_USER_NOT_FOUND = "AUTH_USER_NOT_FOUND",
  AUTH_WRONG_PASSWORD = "AUTH_WRONG_PASSWORD",
  AUTH_TOO_MANY_REQUESTS = "AUTH_TOO_MANY_REQUESTS",
  AUTH_NETWORK_ERROR = "AUTH_NETWORK_ERROR",
  AUTH_REQUIRES_RECENT_LOGIN = "AUTH_REQUIRES_RECENT_LOGIN",
  AUTH_INVALID_CREDENTIAL = "AUTH_INVALID_CREDENTIAL",
  AUTH_USER_DISABLED = "AUTH_USER_DISABLED",
  AUTH_OPERATION_NOT_ALLOWED = "AUTH_OPERATION_NOT_ALLOWED",

  // Profile errors
  PROFILE_NOT_FOUND = "PROFILE_NOT_FOUND",
  PROFILE_ALREADY_EXISTS = "PROFILE_ALREADY_EXISTS",
  PROFILE_UPDATE_FAILED = "PROFILE_UPDATE_FAILED",
  PROFILE_CREATION_FAILED = "PROFILE_CREATION_FAILED",
  PROFILE_UNAUTHORIZED = "PROFILE_UNAUTHORIZED",
  PROFILE_AVATAR_INVALID = "PROFILE_AVATAR_INVALID",
  PROFILE_DISPLAY_NAME_TAKEN = "PROFILE_DISPLAY_NAME_TAKEN",
  PROFILE_DELETE_FAILED = "PROFILE_DELETE_FAILED",

  // Database errors
  DATABASE_CONNECTION_ERROR = "DATABASE_CONNECTION_ERROR",
  DATABASE_PERMISSION_DENIED = "DATABASE_PERMISSION_DENIED",
  DATABASE_QUOTA_EXCEEDED = "DATABASE_QUOTA_EXCEEDED",
  DATABASE_UNAVAILABLE = "DATABASE_UNAVAILABLE",

  // Habitat errors
  HABITAT_NOT_FOUND = "HABITAT_NOT_FOUND",
  HABITAT_ACCESS_DENIED = "HABITAT_ACCESS_DENIED",
  HABITAT_CREATION_FAILED = "HABITAT_CREATION_FAILED",
  HABITAT_UPDATE_FAILED = "HABITAT_UPDATE_FAILED",
  HABITAT_DELETE_FAILED = "HABITAT_DELETE_FAILED",
  HABITAT_JOIN_FAILED = "HABITAT_JOIN_FAILED",
  HABITAT_LEAVE_FAILED = "HABITAT_LEAVE_FAILED",
  HABITAT_ALREADY_MEMBER = "HABITAT_ALREADY_MEMBER",
  HABITAT_NOT_MEMBER = "HABITAT_NOT_MEMBER",

  // Message errors
  MESSAGE_NOT_FOUND = "MESSAGE_NOT_FOUND",
  MESSAGE_SEND_FAILED = "MESSAGE_SEND_FAILED",
  MESSAGE_DELETE_FAILED = "MESSAGE_DELETE_FAILED",
  MESSAGE_UNAUTHORIZED = "MESSAGE_UNAUTHORIZED",

  // Discussion errors
  DISCUSSION_NOT_FOUND = "DISCUSSION_NOT_FOUND",
  DISCUSSION_CREATION_FAILED = "DISCUSSION_CREATION_FAILED",
  DISCUSSION_UPDATE_FAILED = "DISCUSSION_UPDATE_FAILED",
  DISCUSSION_DELETE_FAILED = "DISCUSSION_DELETE_FAILED",
  DISCUSSION_UNAUTHORIZED = "DISCUSSION_UNAUTHORIZED",

  // Poll errors
  POLL_NOT_FOUND = "POLL_NOT_FOUND",
  POLL_CREATION_FAILED = "POLL_CREATION_FAILED",
  POLL_UPDATE_FAILED = "POLL_UPDATE_FAILED",
  POLL_DELETE_FAILED = "POLL_DELETE_FAILED",
  POLL_VOTE_FAILED = "POLL_VOTE_FAILED",
  POLL_UNAUTHORIZED = "POLL_UNAUTHORIZED",

  // Watch Party errors
  WATCH_PARTY_NOT_FOUND = "WATCH_PARTY_NOT_FOUND",
  WATCH_PARTY_CREATION_FAILED = "WATCH_PARTY_CREATION_FAILED",
  WATCH_PARTY_UPDATE_FAILED = "WATCH_PARTY_UPDATE_FAILED",
  WATCH_PARTY_DELETE_FAILED = "WATCH_PARTY_DELETE_FAILED",
  WATCH_PARTY_JOIN_FAILED = "WATCH_PARTY_JOIN_FAILED",
  WATCH_PARTY_LEAVE_FAILED = "WATCH_PARTY_LEAVE_FAILED",
  WATCH_PARTY_FULL = "WATCH_PARTY_FULL",
  WATCH_PARTY_UNAUTHORIZED = "WATCH_PARTY_UNAUTHORIZED",

  // Real-time connection errors
  REALTIME_CONNECTION_FAILED = "REALTIME_CONNECTION_FAILED",
  REALTIME_CHANNEL_ERROR = "REALTIME_CHANNEL_ERROR",
  REALTIME_SUBSCRIPTION_FAILED = "REALTIME_SUBSCRIPTION_FAILED",

  // Generic errors
  UNKNOWN_ERROR = "UNKNOWN_ERROR",
  NETWORK_ERROR = "NETWORK_ERROR",
  RATE_LIMIT_EXCEEDED = "RATE_LIMIT_EXCEEDED",
  TMDB_UNAUTHORIZED = "TMDB_UNAUTHORIZED",
  TMDB_SEARCH_FAILED = "TMDB_SEARCH_FAILED",
}